Move ps4 services to separate module
